XRDDEV-422 refactor, fix permissions
