+++ src/proxy-ui-api/src/main/java/org/niis/xroad/restapi/repository/InternalTlsCertificateRepository.java
@@ -24,28 +24,42 @@
  */
 package org.niis.xroad.restapi.repository;
 
-import ee.ria.xroad.commonui.SignerProxy;
-import ee.ria.xroad.signer.protocol.dto.TokenInfo;
+import ee.ria.xroad.common.util.CryptoUtils;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
+import java.io.FileInputStream;
+import java.io.IOException;
+import java.security.cert.X509Certificate;
 
 /**
- * Repository that handles tokens (acts as a wrapper to SignerProxy)
+ * internal tls certificate repository
  */
 @Slf4j
 @Repository
-public class TokenRepository {
+public class InternalTlsCertificateRepository {
+
+    // as in application_controller.rb
+    private static final String INTERNAL_TLS_CERT_PATH = "/etc/xroad/ssl/internal.crt";
 
     /**
-     * get all tokens
-     * @return
-     * @throws Exception
+     * reads internal tls certificate from file
      */
-    public List<TokenInfo> getTokens() throws Exception {
-        return SignerProxy.getTokens();
+    public X509Certificate getInternalTlsCertificate() {
+        /**
+         *     if File.exists?(INTERNAL_SSL_CERT_PATH)
+         *       File.open(INTERNAL_SSL_CERT_PATH, 'rb') do |f|
+         *         cert = OpenSSL::X509::Certificate.new(f)
+         *       end
+         */
+        try (FileInputStream fileInputStream = new FileInputStream(INTERNAL_TLS_CERT_PATH)) {
+            return CryptoUtils.readCertificate(fileInputStream);
+        } catch (IOException ioe) {
+            log.error("cant read internal tls cert");
+            throw new RuntimeException(ioe);
+        }
     }
+
 
 }
